refactor: Move mix task implementation to Mandate.MixTask module
